requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.8.0
pydantic-core==2.20.0
pydantic-settings==2.6.1
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.27.2
pytest==7.4.3
pytest-asyncio==0.21.1
Jinja2==3.1.2
openai==1.55.3
twilio==8.10.0
galileo-sdk==0.1.0

# Documentation dependencies
mkdocs==1.6.1
mkdocs-material==9.6.16
mkdocs-git-revision-date-localized-plugin==1.4.7